requirements.txt
flask>=2.3
ultralytics>=8.0
opencv-python-headless
Pillow
